Drawing Map in Full
Using nested for loops to count through the segments and polystructs to create a map index.

From this data we can find the tile rotation and index and pass this through to the TIM2Buffer with the map index.

Implemented TIM2Buffer function. Using nested for loops to loop through 32x32 grid. Using GetPixel function to get each pixel and its colour and passing this into SetBufferPixel function to draw the map.

Finally I have began writing a switch statement for the 7 rotation cases to draw each tile with its correct rotation.